+++ api/enums.py
@@ -1,3 +1,4 @@
+from django.utils.translation import gettext_lazy as _
 from djchoices import DjangoChoices, ChoiceItem
 
 
@@ -8,7 +9,7 @@ class TournamentGenderTypes(DjangoChoices):
 
 
 class TeamStateTypes(DjangoChoices):
-    waiting = ChoiceItem()
-    signed_up = ChoiceItem()
-    needs_approval = ChoiceItem()
-    denied = ChoiceItem()
+    waiting = ChoiceItem('waiting', _('waiting'))
+    signed_up = ChoiceItem('signed up', _('signed up'))
+    needs_approval = ChoiceItem('needs approval', _('needs approval'))
+    denied = ChoiceItem('denied', _('denied'))